+++ gulpfile.js
@@ -9,7 +9,6 @@ const Path        = require('path');
 const replace     = require('gulp-replace');
 const sourcemaps  = require("gulp-sourcemaps");
 const spawn       = require('child_process').spawn;
-const through     = require('through2');
 const version     = require('./package.json').version;
 const to5         = require('gulp-6to5');
 
@@ -25,10 +24,7 @@ gulp.task('default', ['build'], function() {
 
 // Compile ES6 in src to ES5 in lib
 gulp.task('build', ['clean'], function() {
-  const options = {
-    // Support iterators without polyfill
-    optional:     ['coreAliasing']
-  };
+  const options = require('./src/6to5.json');
   const compile = gulp.src('src/**/*.js')
     .pipe(sourcemaps.init())
     .pipe(to5(options))