+++ test/es6.js
@@ -1,6 +1,3 @@
 const to5 = require('6to5/register');
 
-to5({
-  // Support async/await
-  experimental: true
-});
+to5(require('../src/6to5.json'));